Convert enableDraw() method to a new 'edit' behavior.
